backend/requirements.txt
functions-framework==3.5.0
google-generativeai==0.3.1
flask==3.1.2